Initial sync of the MIDP on CDC work in progress branch. Contents include AMS refactoring, CDC platform support for WINCE and various stubbing eforts to disconnect subsystems that have not been refactored yet for integration into the trunk.
Also merged changes from the trunk added since the
initial branch was created.
(svn merge -r419:523 \
 https://phoneme.dev.java.net/svn/phoneme/components/midp/trunk)

This branch will be updated periodically while 
the MIDP on CDC high level abstraction interfaces 
are introduced into the shared code. Groups of 
files will be migrated to the trunk as they are 
reviewed and test for both CDC and CLDC stacks.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/branches/midp-mr2-ext-cdc@540 6dfe35fe-931f-0410-af5f-a91b034811cd