CR 6515922, Wrong Canvas repaint for an empty area, TCK failure.
Entire Canvas was painted on repaint request for an empty area.
It happened because of no check that body layer is not dirty, so 
entire body layer was repainted. The checks are added. Canvas is 
optimized also to not post repaint event on empty area request.

Reviewed by Svetlana Gurenkova, thanks!

NOTE: One more problem is detected, it will be continued under new CR.
The optimized Canvas painting should not be done when the Canvas is
overlapped with a visible higher layer.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@1444 6dfe35fe-931f-0410-af5f-a91b034811cd