CR 6513537: autotest doesn't allow the permissions absent from the suite's jad file
An additional optional parameter for AutoTester was introduced that may  follow the name of domain
after the colon - a list of permissions that must be allowed even if they are missing from the jad, for example:

./autotest_g http://<host>/midlets/HelloMIDlet.jad operator:all

NOTE: work on this CR is not finished after this putback. The more consistent solution is going to be integrated:
- remove Installer.setExtraPermissions().
- write an autotester that first did an HTTP download of the JAD and JAR to a file,
then added attributes to the JAD and Manifest, then do a file install.

svn merge -r1166:1489 ../branches/midp-mr2-cr-6513537

Webrev at:
http://spiro.sfbay/jcse/wireless/leapfrog/webrev/zav/midp-mr2-cr-6513537/
Reviewed by Stephen Flores and Sergey Troshin, thanks.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@1491 6dfe35fe-931f-0410-af5f-a91b034811cd