CR 6516980: Midlet is inaccessible if it doesn't use the screen in case of the security alert
The problem is that in the current implementation the midlet is sent in foreground on the "midlet activated" event.
It happens after exiting from the midlet's startApp() method, so if it is required to show an alert *before* startApp()
is finished, it is not shown because the midlet still be in the background.

An attempt to send the midlet into foreground from the AMS also fails because the midlet has no display
at this moment.

After this fix a display for the midlet is created at the same time as the midlet (in the midlet's constructor).

svn merge -r1940:2017 ../branches/midp-mr2-cr-6516980

Webrev at:
http://spiro.sfbay/jcse/wireless/leapfrog/webrev/zav/midp-mr2-cr-6516980/ 
Reviewed by Sergey Nazarkin, Alexey Kornev and Kristina Kudrjavceva, thanks.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@2058 6dfe35fe-931f-0410-af5f-a91b034811cd