Merge branch midp-mr2-cr-6539336-2:
New fix for 6539336. Instead of creating the
separate class zip file using CDC classes, we
compile the 13 MIDP classes separately and
create the zip file using them. This is because
CDC Timer class has an extra public constructor
that's not in MIDP spec, so we can't use CDC
classes for the MIDP signature test. Reviewed
by Gary.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@4237 6dfe35fe-931f-0410-af5f-a91b034811cd