CR 6535880 (JSR226: com/sun/jsr226/functional/m2d/Render#testTransparency - incorrect transparency).
The problem is in MIDP gx_draw_rgb function, premulOpacity function is OK.
gx_draw_rgb uses correct algorithm (please see description for AlphaComposite class in J2SE, and specifically SRC_OVER composition).
Anyway it utilizes incorrect operations for '*' and '+'  (it uses bit operations for this purpose).
Currently used formula is: composition = (xA & value) | (XAInv & background).

This can lead to strange visual effects, for instance, the artifact appears on gray color circle in the test.
Let 'value' in ARGB format is 0x66808080 (gray with 0.4 transparency) and 'background' is 0xffffffff (opaque white).

While transparency is changed in the range from 0 to 0x7f, 'xA*value' is always 0:
0x80 == 1000000b while 0...0x7f == 00000000...01111111, i.e. bit multiplication of these numbers is always 0.

As result, a color component of composition is changed linear from 0xff to 0x80 only due 'XAInv*background' is changed.
When transparency is 0.5, 'xA*value' is immediately changed from 0 to 0x80, as result, composition immediately rises from 0x80 to 0xff (it's easy to see sharp color alternation in this case). After that, composition starts decreasing again from 0xff to 0x80.

I.e. the composition appearance in [0,0.5) range is the same as in [0.5, 1.0].

Merged -r4495:5843 with midp-cr-6535880.

Reviewed by Alexey Zavitaev.

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@5844 6dfe35fe-931f-0410-af5f-a91b034811cd