CR 6594148: Invalid scrolling of form items in UIDemo
The problem was with long choice group widget. When the focus was outside of it,
the pointer event at some item of the choice group caused scrolling to the top of
the choice group. It confused user since selected item could disappear in this way,
and scroll bar showed invalid position. 

The fix enables selected item to be visible partially and to not scroll entire 
choice group to its top.

Reviewed by Olga & Kristina, thanks.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@7042 6dfe35fe-931f-0410-af5f-a91b034811cd