In the case input mode has own Displayable (e.g. symbol input from the table), screen change event is posted and processed by the system. On return from input mode with Displayable, the previous Displayable is restored and text field component is notified of input mode changing.
This scenario caused the problem since when text field component is notified about 
input mode changing, its current display is null. It happened because screen change 
event is posted to the system but not processed yet.

To fix the problem the cached input session value is used, and this decision 
limits text edit field to be associated with one Display only. Proper IMPL_NOTE
is provided. Cached input session instance gets minor benefit in performance also.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/branches/midp-mr3-cr-6605277@7364 6dfe35fe-931f-0410-af5f-a91b034811cd