6658980: cdc should support builds where part of the port is in a separate component
More changes for 6658980 that are needed to support midp builds when the OS port is in another build directory. The need for these changes was discovered when testing out wm5 builds when the win32 and win32-arm port was moved to another component.

The main changes were to export TARGET_OS to the midp build, and all make all the CDC_XXX_COMPONENT_DIR variables exported so midp can access them. Also, I removed the redundant passing of USE_SSL=$(USE_SLL) to the midp build (it was done twice).

See r9524 in the cdc component for details on the fix for 6658980.

Tested with wm5 and win32/x86 builds, both from the cdc component and from another component.

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@9534 6dfe35fe-931f-0410-af5f-a91b034811cd