MIDP likes to set its own compiler flags. It also imports them, which can result in conficts. For CDC, MIDP imports all the CDC CFLAGS from the CCFLAGS_SPEED, which is added to EXTRA_CFLAGS. It get CCFLAGS_SPEED from build_defs.mk, which is generated during CDC build.
The problem I've ran into is that sometimes the flags that MIDP wants to set are in conflict with the ones CDC passes in, and generates warnings. A common example is when building with CVM_OPTIMIZED=true CVM_DEBUG=true. USE_DEBUG=true is passed to MIDP since CVM_DEBUG=true. You end up with CDC passing in /O2 /Ob2 /Ot because CVM_OPTIMIZED=true. MIDP is adding /Od because USE_DEBUG=true, which is in conflict with the other optimization options and generates a warning. This problem stems from the fact that USE_DEBUG=true is a loaded option in MIDP. It means:

    * Don't optimize code
    * Generate debugger symbols
    * Compile in extra debugging code

CDC controls these three build options separately with CVM_OPTIMIZED, CVM_SYMBOLS, and CVM_DEBUG.

To avoid this conflict this commit changes it so MIDP does set any compiler flags that CDC already passes in CCFLAGS_SPEED. Note -DWIN32 and -D_WINDOWS are already present in CVM_DEFINES, which MIDP also imports and adds to EXTRA_CFLAGS.

This change was approved a couple of months ago in an email discussion that included the cdc and midp teams.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@13058 6dfe35fe-931f-0410-af5f-a91b034811cd