CR 6768385: Need a way to specify parameters to runMidlet on win32_emul platform
A new library ams_params was introduced.
It has two implementations: one reads the parameters for runMidlet
from "ams_parameters.txt" file (must be located in the same directory
with the runMidlet) and is used in Javacall builds; the second one is
stubbed out and is used in all other builds.

svn merge -r15210:15239 ../branches/midp-cr-6768385

Reviewed by Kristina Kudrjavceva, Michael Gasanenko.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@15240 6dfe35fe-931f-0410-af5f-a91b034811cd