CR 6661462: Cannot recover from the errors when _suites.dat is corrupted
When the AMS is starting, checkSuitesIntegrity() is called to ensure that the suite database
is not corrupted. If it is, an attempt to repair it is taken. If it was successful, the user
is notified that the problem existed and then the AMS continues to load. If the recovery
attempt has failed, the appropriate error message is displayed and then SJWC exits.

svn merge -r15368:15438 ../branches/midp-cr-6661462

Reviewed by Kristina Kudrjavceva.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@15439 6dfe35fe-931f-0410-af5f-a91b034811cd