CR 6776964: Regression: a lot of MIDP and MMAPI TCK tests fail with SecurityException
The root cause of the problem is that both .../http/Protocol and .../https/Protocol classes
contain private field "permissionChecked", but https/Protocol is derived from http/Protocol.
So if the permission was checked in the https/Protocol class, "permissionChecked" field of
the http/Protocol class still be false. Thus SecurityException is thrown during an attempt
to call any method of http/Protocol.

svn merge -r15797:15804 ../branches/midp-cr-6776964

Reviewed by Leonid Romanov.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@15805 6dfe35fe-931f-0410-af5f-a91b034811cd