CR 6776285 (Impossible to run several MIDlets from the same suite if one was run several times)
The issue is in the MIDletSelector.commandAction method.

There is a guard there to protect from launching a MIDlet that is being launched already.
It checks whether "selectedMidlet != -1", and if yes, just ignores the command.

A bug in the commandAction method's code was introduced by the code
that detects that if the MIDlet (that user selected from the list) is already 
running then bring it to the foreground instead of launching it.
But the code doesn't reset selectedMIDlet to "-1" 
so all further "Launch" and "Bring to the foreground" commands are ignored.

Merged as 15837:15847 with midp-cr-6776285.

Reviewed by Sergey Troshin and Alexey Zavitaev.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/midp/trunk@15849 6dfe35fe-931f-0410-af5f-a91b034811cd